+++ src/redux/config/firebaseConfig.js
